refactoring tpr/icons and adding 2 new components SpinnerLoadingCircle and SpinnerLoadingProgress
