+++ packages/icons/src/components/spinners/spinners.module.scss
@@ -1,0 +1,143 @@
+@import '@tpr/theming/lib/variables.scss';
+
+.wrapper1 {
+	display: inline-block;
+	text-align: center;
+
+	.spinner1 {
+		position: relative;
+		width: 80px;
+		height: 80px;
+		overflow: visible;
+		margin: 0 auto;
+	}
+	.spinner1 div {
+		position: absolute;
+		border: 10px solid #cdd500;
+		opacity: 1;
+		border-radius: 50%;
+		//animation: lds-ripple 1s cubic-bezier(0, 0.2, 0.8, 1) infinite;
+		animation: lds-ripple 1s cubic-bezier(0, 0.9, 0.9, 1) infinite;
+	}
+	.spinner1 div:nth-child(2) {
+		animation-delay: -0.5s;
+		border: 10px solid #cdd500;
+	}
+
+	.text {
+		font-family: $fonts-sans-serif;
+		font-size: 16px;
+		color: $black;
+		text-align: center;
+	}
+}
+
+@keyframes lds-ripple {
+	0% {
+		top: 36px;
+		left: 36px;
+		width: 0px;
+		height: 0;
+		opacity: 1;
+	}
+	100% {
+		top: 0px;
+		left: 0px;
+		width: 72px;
+		height: 72px;
+		opacity: 0;
+	}
+}
+
+/* ----------------------------------- */
+
+.wrapper2 {
+	display: inline-block;
+	text-align: center;
+
+	.spinner2 {
+		display: inline-block;
+		position: relative;
+		width: 80px;
+		height: 80px;
+	}
+	.spinner2 div {
+		transform-origin: 40px 40px;
+		animation: lds-spinner 1.2s linear infinite;
+	}
+	.spinner2 div:after {
+		content: ' ';
+		display: block;
+		position: absolute;
+		top: 3px;
+		left: 37px;
+		width: 6px;
+		height: 18px;
+		border-radius: 20%;
+		background: #483b8b;
+	}
+	.spinner2 div:nth-child(1) {
+		transform: rotate(0deg);
+		animation-delay: -1.1s;
+	}
+	.spinner2 div:nth-child(2) {
+		transform: rotate(30deg);
+		animation-delay: -1s;
+	}
+	.spinner2 div:nth-child(3) {
+		transform: rotate(60deg);
+		animation-delay: -0.9s;
+	}
+	.spinner2 div:nth-child(4) {
+		transform: rotate(90deg);
+		animation-delay: -0.8s;
+	}
+	.spinner2 div:nth-child(5) {
+		transform: rotate(120deg);
+		animation-delay: -0.7s;
+	}
+	.spinner2 div:nth-child(6) {
+		transform: rotate(150deg);
+		animation-delay: -0.6s;
+	}
+	.spinner2 div:nth-child(7) {
+		transform: rotate(180deg);
+		animation-delay: -0.5s;
+	}
+	.spinner2 div:nth-child(8) {
+		transform: rotate(210deg);
+		animation-delay: -0.4s;
+	}
+	.spinner2 div:nth-child(9) {
+		transform: rotate(240deg);
+		animation-delay: -0.3s;
+	}
+	.spinner2 div:nth-child(10) {
+		transform: rotate(270deg);
+		animation-delay: -0.2s;
+	}
+	.spinner2 div:nth-child(11) {
+		transform: rotate(300deg);
+		animation-delay: -0.1s;
+	}
+	.spinner2 div:nth-child(12) {
+		transform: rotate(330deg);
+		animation-delay: 0s;
+	}
+
+	.text {
+		font-family: $fonts-sans-serif;
+		font-size: 16px;
+		color: $black;
+		text-align: center;
+	}
+}
+
+@keyframes lds-spinner {
+	0% {
+		opacity: 1;
+	}
+	100% {
+		opacity: 0;
+	}
+}
