--- packages/icons/src/icons.mdx
@@ -1,101 +1,0 @@
----
-name: Icons
-route: /icons
----
-
-import { Props } from 'docz';
-import { Playground } from '@playground';
-import {
-	CheckboxChecked,
-	CheckboxBlank,
-	RadioButtonChecked,
-	RadioButtonUnchecked,
-	UnfoldMore,
-	CheckedCircle,
-	ErrorCircle,
-	ArrowUp,
-	ArrowDown,
-	ArrowLeft,
-	ArrowRight,
-	Cross,
-} from './icons';
-
-# Icons
-
-Icons for tpr apps
-
-## Usage
-
-```js
-import { CheckboxChecked, CheckboxBlank, UnfoldMore } from '@tpr/icons';
-```
-
-| Components           |
-| -------------------- |
-| CheckboxChecked      |
-| CheckboxBlank        |
-| RadioButtonChecked   |
-| RadioButtonUnchecked |
-| UnfoldMore           |
-
-## Checkbox
-
-<Playground>
-	<CheckboxChecked />
-	<CheckboxBlank />
-</Playground>
-
-## Radio Button
-
-<Playground>
-	<RadioButtonChecked />
-	<RadioButtonUnchecked />
-</Playground>
-
-## UnfoldMore
-
-<Playground>
-	<UnfoldMore />
-</Playground>
-
-## Checked Circle
-
-<Playground>
-	<CheckedCircle fill="#1e7e34" />
-</Playground>
-
-## Error Circle
-
-<Playground>
-	<ErrorCircle fill="#a91717" />
-</Playground>
-
-## Arrow Up
-
-<Playground>
-	<ArrowUp />
-</Playground>
-
-## Arrow Down
-
-<Playground>
-	<ArrowDown />
-</Playground>
-
-## Arrow Left
-
-<Playground>
-	<ArrowLeft />
-</Playground>
-
-## Arrow Right
-
-<Playground>
-	<ArrowRight />
-</Playground>
-
-## Cross
-
-<Playground>
-	<Cross />
-</Playground>
