+++ packages/icons/src/index.mdx
@@ -1,0 +1,148 @@
+---
+name: Icons
+route: /icons
+---
+
+import { Props } from 'docz';
+import { Playground } from '@playground';
+import {
+	CheckboxChecked,
+	CheckboxBlank,
+	RadioButtonChecked,
+	RadioButtonUnchecked,
+	UnfoldMore,
+	CheckedCircle,
+	ErrorCircle,
+	ArrowUp,
+	ArrowDown,
+	ArrowLeft,
+	ArrowRight,
+	Cross,
+	LoadingSpinnerCircle,
+	LoadingSpinnerProgress,
+} from './index';
+
+# Icons
+
+Icons for tpr apps
+
+## Usage
+
+```js
+import { CheckboxChecked, CheckboxBlank, UnfoldMore } from '@tpr/icons';
+```
+
+| Components             |
+| ---------------------- |
+| CheckboxChecked        |
+| CheckboxBlank          |
+| RadioButtonChecked     |
+| RadioButtonUnchecked   |
+| UnfoldMore             |
+| CheckedCircle          |
+| ErrorCircle            |
+| ArrowUp                |
+| ArrowDown              |
+| ArrowLeft              |
+| ArrowRight             |
+| Cross                  |
+| LoadingSpinnerCircle   |
+| LoadingSpinnerProgress |
+
+## Checkbox
+
+<Playground>
+	<CheckboxChecked />
+	<CheckboxBlank />
+</Playground>
+
+## Radio Button
+
+<Playground>
+	<RadioButtonChecked />
+	<RadioButtonUnchecked />
+</Playground>
+
+## UnfoldMore
+
+<Playground>
+	<UnfoldMore />
+</Playground>
+
+## Checked Circle
+
+<Playground>
+	<CheckedCircle fill="#1e7e34" />
+</Playground>
+
+## Error Circle
+
+<Playground>
+	<ErrorCircle fill="#a91717" />
+</Playground>
+
+## Arrow Up
+
+<Playground>
+	<ArrowUp />
+</Playground>
+
+## Arrow Down
+
+<Playground>
+	<ArrowDown />
+</Playground>
+
+## Arrow Left
+
+<Playground>
+	<ArrowLeft />
+</Playground>
+
+## Arrow Right
+
+<Playground>
+	<ArrowRight />
+</Playground>
+
+## Cross
+
+<Playground>
+	<Cross />
+</Playground>
+
+## LoadingSpinnerCircle
+
+<Playground>
+	<LoadingSpinnerCircle />
+</Playground>
+
+## LoadingSpinnerCircle icon only
+
+<Playground>
+	<LoadingSpinnerCircle iconOnly={true} />
+</Playground>
+
+## LoadingSpinnerCircle with custom text
+
+<Playground>
+	<LoadingSpinnerCircle text="Please wait while the app is loading ..." />
+</Playground>
+
+## LoadingSpinnerProgress
+
+<Playground>
+	<LoadingSpinnerProgress />
+</Playground>
+
+## LoadingSpinnerProgress icon only
+
+<Playground>
+	<LoadingSpinnerProgress iconOnly={true} />
+</Playground>
+
+## LoadingSpinnerProgress with custom text
+
+<Playground>
+	<LoadingSpinnerProgress text="Please wait while the app is loading ..." />
+</Playground>
