+++ packages/icons/src/index.ts
@@ -1,1 +1,7 @@
-export * from './icons';
+export * from './components/types';
+export * from './components/global';
+export * from './components/arrows/arrows';
+export * from './components/circle/circle';
+export * from './components/cross/cross';
+export * from './components/form/form';
+export * from './components/spinners/spinners';
